Incorporated tables joins
Instead of using excel to prepare the data (getting parent id, course, and parent name), I added the parent info into the sql database so we could utilize joins. This will be faster and easier than preparing the dat in excel the importing the data into the data base.

Also changed the having clause to greater or equal (>=) to identify which parent took multiple classes. Identified an issue with finding parents who don't meet all conditions in the "IN" statement, but have taken multiples of one class meeting the condition in the having statement.

I.e. The LinkedIn certificate requires 2 classes  (LinkedIn-1 and LinkedIn-2), but if a parent took LinkedIn-1 multiple times they would meet the condition for earning the certificate.