Having statement issue; no data pulled
Found an issue with expanding the list of certificates and the required classes. The HAVING clause works when the table only contains the classes required to earn a certificate.  However, when these certificates are in multiple columns in the same table, then the query pulls zero records. I assume its because its taking the count of other columns from the other certificates. 

I will need to test column counts columns in each table.