Added additional logic on how code works
The having statement is counts all classes, but is set at the predetermined number. When putting just the count aggregate into the select statement it counts all of the classes a student has taken.

The way the where statement works is it's specifying what classes to pull, and the having statement specifies the number of classes  required to earn a certificate. The having statement uses distinct count to filter out students who repeated a class multiple times. Together they work together to identify which parents earn a certificate.