Test Code for Dates
This block of code is used to try to identify how DBeaver's SQLite stores dates. I had trouble getting the where condition to store dates when I set the preferred date for mate to MM/DD/YYYY. It turns on that was only for displaying data and not storing data. 
With this I now know I can import dates if they're stored as YYYY-MM-DD.