reimported and added additional clean up queries
The dates were imported improperly. Dates needed to be in YYYY-MM-DD to properly function as dates in DBeaver. Also added a function to trim and capitalize the first letter of each word after a space. Added in a query to clean Phoenix high school. Added the GROUP BY function so aggregated list display values that were missing before.